LUQ Stream Chemistry Data for Quebrada Sonadora (QS) site

Data include stream chemistry data for the Quebrada Sonadora (QS) location part of the Luqillo LTER site

luq_streamchem

Format

A tibble with 597 rows and 18 variables:

Sample_ID

a character denoting Watershed or Collector's location where the sample is taken

Sample_Date

a date denoting the date when the sample was collected

Gage_Ht

a number denoting the height of the water surface above the gage datum (zero point) at sampling time (in meters)

Temp

a number denoting the temperature of the sample collected

pH

a number pH (Electrometric with combination electrode)

Cond

a number denoting the conductivity (Conductivity bridge in micro siemens/cm)

Cl

a number denoting the concentration of Chloride (up to 1989 by automated colorimetric analysis (thiocyanate); after 1989 by liquid chromatography with conductivity detection) in mg Cl/L

NO3

a number denoting the concentration of dissolved nitrate in the stream water in ug N/L

SO4-S

a number denoting the concentration of Sulfate (Liquid chromatography; conductivity detection) in mg/L

Na

a number denoting the concentration of Sodium (up to 1994 by atomic absorption spectroscopy) in mg/L

K

a number denoting the concentration of Potassium (up to 1994 by atomic absorption spectroscopy) in mg/L

Mg

a number denoting the concentration of Magnesium (up to 1994 by atomic absorption spectroscopy) in mg/L

Ca

a number denoting the concentration of Calcium (up to 1994 by atomic absorption spectroscopy) in mg/L

NH4-N

a number denoting the concentration of Ammonium Nitrogen (automated colorimetric analysis (phenolhypochlorite)) in ug N/L

DOC

a number denoting Dissolved Organic Carbon in the stream water ion milligrams carbon per liter (mg C/L)

DIC

a number denoting Dissolved Inorganic Carbon (by syringe stripping and infrared spectroscopy (Stainton 1973)) in mg C/L

SiO2

a number denoting the concentration of Silica (automated colorimetric analysis (molybdate blue)) in mg SiO2/L

TSS

a number denoting the Total Suspended Sediments in mg/L

Source

McDowell, W. 2021. Chemistry of stream water from the Luquillo Mountains ver 4923056. Environmental Data Initiative. https://doi.org/10.6073/pasta/0a09f5aa2e6f11451553c92b102279a6